Fix: MCP Server: MCP Inspector run fixed to load the server appropriately. Change: MCP Server: Enhance server startup (python mcp_server.py) with new configuration methods and output formatting. THose method were moved from mac_server_startup.sh to mcp_server.py. Change: MCP Server: bash script does "ln -s ../server/genericsuite_codegen ." to link the server to the root of the project, instead of copying it. Change: MCP Server: Update Makefile to include requirements export command. Change: MCP Server: Update package.json for improved keyword formatting. GS-172
