+++ mcp-server/Makefile
@@ -19,6 +19,9 @@ help:
 install:
 	poetry install
 
+requirements:
+	poetry export -f requirements.txt --output requirements.txt --without-hashes
+
 # Check environment configuration
 show-env-vars:
 	@echo "Checking environment configuration..."
