+++ server/genericsuite_codegen/mcp_server/__init__.py
@@ -6,6 +6,24 @@ for the GenericSuite CodeGen system, exposing AI agent capabilities
 as standardized MCP tools and resources.
 """
 
-from .server import create_mcp_server, MCPConfig
+from .server import (
+    create_mcp_server,
+    MCPConfig,
+    load_environment,
+    validate_environment,
+    get_mcp_config,
+    report_mcp_config,
+    print_output,
+    logger,
+)
 
-__all__ = ["create_mcp_server", "MCPConfig"]
+__all__ = [
+    "create_mcp_server",
+    "MCPConfig",
+    "load_environment",
+    "validate_environment",
+    "get_mcp_config",
+    "report_mcp_config",
+    "print_output",
+    "logger",
+]
